Changes to get drishti to compile

+++ drishti/keyframeinformation.cpp
@@ -937,7 +937,7 @@ KeyFrameInformation::save(fstream &fout)
   sprintf(keyword, "rotation");
   fout.write((char*)keyword, strlen(keyword)+1);
   Vec axis;
-  float angle;
+  double angle;
   m_rotation.getAxisAngle(axis, angle);
   f[0] = axis.x;
   f[1] = axis.y;
